Fix error: Each class must be in a file by itself
