+++ public/app/themes/clarity/functions.php
@@ -52,7 +52,8 @@ require_once 'inc/admin/wp-admin-bar.php';
 
 require_once 'inc/about-us.php';
 require_once 'inc/acf.php';
-require_once 'inc/amazon-s3-and-cloudfront.php';
+require_once 'inc/amazon-s3-and-cloudfront-for-minio.php';
+require_once 'inc/amazon-s3-and-cloudfront-signing.php';
 
 require_once 'inc/api/get-posts-rest-api.php';
 require_once 'inc/api/campaign-api.php';
